modify: add endpoint for app
